Proposal for a way for sources to sign articles' media (SCS #48)
This is necessary since otherwise there's no warranty that a malicious third party won't MITM and replace media, or that a malicious node won't intentionally serve the wrong media. This can get particularly dangerous, e.g. if a government tries to replace communication material from an opposing party with dangerous or misleading reproductions.